Refactor signature algorithms into own namespace
