+++ src/SimpleJWT/Crypt/AlgorithmFactory.php
@@ -48,10 +48,10 @@ class AlgorithmFactory {
     /** @var array<string, string> $alg_map */
     static $alg_map = [
         // Signature algorithms
-        '/^ES\d+$/' => 'SimpleJWT\Crypt\OpenSSLSig',
-        '/^ES256K$/' => 'SimpleJWT\Crypt\OpenSSLSig',
-        '/^RS\d+$/' => 'SimpleJWT\Crypt\OpenSSLSig',
-        '/^HS\d+$/' => 'SimpleJWT\Crypt\HMAC',
+        '/^ES\d+$/' => 'SimpleJWT\Crypt\Signature\OpenSSLSig',
+        '/^ES256K$/' => 'SimpleJWT\Crypt\Signature\OpenSSLSig',
+        '/^RS\d+$/' => 'SimpleJWT\Crypt\Signature\OpenSSLSig',
+        '/^HS\d+$/' => 'SimpleJWT\Crypt\Signature\HMAC',
 
         // Key management algorithms (derivation or encryption)
         '/^dir$/' => 'SimpleJWT\Crypt\DirectEncryption',
@@ -70,7 +70,7 @@ class AlgorithmFactory {
 
     /** @var array<string, string> $use_map */
     private static $use_map = [
-        Algorithm::SIGNATURE_ALGORITHM => 'SimpleJWT\Crypt\SignatureAlgorithm',
+        Algorithm::SIGNATURE_ALGORITHM => 'SimpleJWT\Crypt\Signature\SignatureAlgorithm',
         Algorithm::ENCRYPTION_ALGORITHM => 'SimpleJWT\Crypt\Encryption\EncryptionAlgorithm',
         Algorithm::KEY_ALGORITHM => 'SimpleJWT\Crypt\KeyManagementAlgorithm'
     ];
@@ -141,7 +141,7 @@ class AlgorithmFactory {
      * @return void
      */
     static public function addNoneAlg() {
-        self::$alg_map['/^none$/'] = 'SimpleJWT\Crypt\None';
+        self::$alg_map['/^none$/'] = 'SimpleJWT\Crypt\Signature\None';
     }
 
     /**
