+++ src/SimpleJWT/Crypt/Signature/HMAC.php
@@ -33,7 +33,7 @@
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-namespace SimpleJWT\Crypt;
+namespace SimpleJWT\Crypt\Signature;
 
 use SimpleJWT\Keys\KeyException;
 use SimpleJWT\Util\Util;
