+++ src/SimpleJWT/Crypt/Signature/OpenSSLSig.php
@@ -33,8 +33,9 @@
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-namespace SimpleJWT\Crypt;
+namespace SimpleJWT\Crypt\Signature;
 
+use SimpleJWT\Crypt\CryptException;
 use SimpleJWT\Keys\Key;
 use SimpleJWT\Keys\KeyException;
 use SimpleJWT\Util\ASN1;
