+++ src/SimpleJWT/Crypt/Signature/SHA2.php
@@ -33,51 +33,33 @@
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-namespace SimpleJWT\Crypt;
+namespace SimpleJWT\Crypt\Signature;
+
+use SimpleJWT\Crypt\Algorithm;
+use SimpleJWT\Util\Util;
 
 /**
- * Implements the `none` signature algorithm.
- * 
- * Note that the `none` algorithm should only be used in JWTs which
- * are cryptographically protected by other means.
- * 
- * By default, the `none` algorithm is disabled in SimpleJWT.  Attempts
- * to decode a JWT with a `none` algorithm will return a
- * {@link SimpleJWT\InvalidTokenException}.  In order to enable this
- * algorithm, call {@link AlgorithmFactory::addNoneAlg()} static
- * method.
- * 
- * @link https://datatracker.ietf.org/doc/html/rfc8725.html#section-3.2
- * @codeCoverageIgnore
+ * Abstract class for SHA2-based signature algorithms.
  */
-class None extends Algorithm implements SignatureAlgorithm {
-    public function __construct($alg) {
+abstract class SHA2 extends Algorithm implements SignatureAlgorithm {
+    /** @var int the length, in bits, of the SHA-2 hash */
+    protected $size;
+
+    /**
+     * Creates an instance of this algorithm.
+     * 
+     * @param ?string $alg the algorithm
+     * @param ?int $size the length, in bits, of the SHA-2 hash
+     */
+    protected function __construct($alg, $size) {
         parent::__construct($alg);
-    }
-
-    public function getKeyCriteria() {
-        return [];
-    }
-
-    public function getSupportedAlgs() {
-        return ['none'];
-    }
-
-    public function sign($data, $keys, $kid = null) {
-        return '';
+        $this->size = $size;
     }
 
     public function shortHash($data) {
-        return '';
-    }
-
-    public function verify($signature, $data, $keys, $kid = null) {
-        if ($kid != null) return false;
-        return ($signature === '');
-    }
-
-    public function getSigningKey($keys, $kid = null) {
-        return null;
+        $hash = hash('sha' . $this->size, $data, true);
+        $short = substr($hash, 0, $this->size / 16);
+        return Util::base64url_encode($short);
     }
 }
