+++ src/SimpleJWT/Crypt/Signature/SignatureAlgorithm.php
@@ -33,8 +33,9 @@
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-namespace SimpleJWT\Crypt;
+namespace SimpleJWT\Crypt\Signature;
 
+use SimpleJWT\Crypt\CryptException;
 use SimpleJWT\Keys\Key;
 use SimpleJWT\Keys\KeySet;
 use SimpleJWT\Keys\KeyException;
